Downloads/Shaman.io/datasetforge/requirements.txt
streamlit>=1.32.0
httpx
spacy
numpy
json-repair
pypdf
python-docx
chardet
scikit-learn
